requirements.txt
langchain_openai
langchain_core
langchain-ollama
python-dotenv
streamlit
langchain_community
langserve